// src/models/ComandoAdministracionConsultas.ts
export class ComandoAdministracionConsultas {
  cuil: string;
  numero: number;
  idEstado: number;
  solicitud: string;
  usuario: string;
  idAsunto: number;
  idArea: number;
  idOficina: number;
  fechaDesde: string;
  fechaHasta: string;
  consultasMarcadas: boolean;
  consultasCerradas: boolean;
}